+++ python/samples/02-agents/providers/foundry/foundry_chat_client_with_toolbox.py
@@ -2,52 +2,48 @@
 
 import asyncio
 import os
+from collections.abc import Callable
+from typing import Any
 
-from agent_framework import Agent
-from agent_framework.foundry import FoundryChatClient, select_toolbox_tools
-from azure.identity import AzureCliCredential
+from agent_framework import Agent, MCPStreamableHTTPTool
+from agent_framework.foundry import FoundryChatClient
+from azure.core.credentials import TokenCredential
+from azure.identity import AzureCliCredential, DefaultAzureCredential, get_bearer_token_provider
 from dotenv import load_dotenv
 
 # Load environment variables from .env file
 load_dotenv()
 
 """
-Foundry Chat Client with Toolbox Example
+Foundry Toolbox via MAF ``MCPStreamableHTTPTool``
 
-This sample demonstrates loading a named, versioned Foundry toolbox into an
-Agent via ``FoundryChatClient.get_toolbox()``. A toolbox is a server-side
-bundle of tool configurations (code interpreter, file search, MCP, web search,
-etc.) configured in the Foundry portal or via the raw SDK.
+Instead of fetching the toolbox and fanning out individual tool specs, point
+MAF's ``MCPStreamableHTTPTool`` at the toolbox's MCP endpoint. The agent
+discovers and calls the toolbox's tools over MCP at runtime.
 
 Prerequisites:
-- A Microsoft Foundry project
-- A toolbox already configured in that project (set TOOLBOX_NAME below)
+- A Microsoft Foundry project with a toolbox configured
 - FOUNDRY_PROJECT_ENDPOINT and FOUNDRY_MODEL environment variables set
+- FOUNDRY_TOOLBOX_ENDPOINT: the toolbox's MCP endpoint URL, e.g.
+  ``https://<account>.services.ai.azure.com/api/projects/<project>/toolsets/<name>/mcp?api-version=v1``
+- Azure CLI authentication (``az login``)
 """
 
-# Replace with your own Foundry toolbox name and version.
+# Must match the ``<name>`` segment of FOUNDRY_TOOLBOX_ENDPOINT.
 TOOLBOX_NAME = "research_toolbox"
-TOOLBOX_VERSION = "1"
-# Used only by combine_toolboxes() — swap in a second toolbox you own.
-SECOND_TOOLBOX_NAME = "analysis_toolbox"
-SECOND_TOOLBOX_VERSION = "1"
-
-# Replace with any question that exercises the tools configured in your toolbox.
-QUERY = "Introduce yourself and briefly describe the tools you can use to help me."
 
 
 def create_sample_toolbox(name: str) -> str:
     """Create (or replace) a toolbox version in the Foundry project.
 
     Toolboxes are normally configured in the Foundry portal or a deployment
-    script, not the application itself. This helper exists so the samples can
+    script, not the application itself. This helper exists so the sample can
     be run end-to-end without first setting a toolbox up by hand — delete any
-    existing toolbox under ``name``, then create a fresh version containing an
-    MCP tool, a web search tool, and a code interpreter tool. Returns the
-    created version identifier.
+    existing toolbox under ``name``, then create a fresh version containing a
+    single MCP tool. Returns the created version identifier.
     """
     from azure.ai.projects import AIProjectClient
-    from azure.ai.projects.models import CodeInterpreterTool, MCPTool, Tool, WebSearchTool
+    from azure.ai.projects.models import MCPTool, Tool
     from azure.core.exceptions import ResourceNotFoundError
 
     with (
@@ -68,9 +64,6 @@ def create_sample_toolbox(name: str) -> str:
             )
         ]
 
-        tools.append(WebSearchTool(name="web_search"))
-        tools.append(CodeInterpreterTool(name="code_interpreter"))
-
         created = project_client.beta.toolboxes.create_version(
             name=name,
             description="Toolbox version with MCP require_approval set to 'never'.",
@@ -80,99 +73,46 @@ def create_sample_toolbox(name: str) -> str:
         return created.version
 
 
+def make_toolbox_header_provider(credential: TokenCredential) -> Callable[[dict[str, Any]], dict[str, str]]:
+    """Build a header_provider that injects a fresh Azure AI bearer token on every MCP request."""
+    get_token = get_bearer_token_provider(credential, "https://ai.azure.com/.default")
+
+    def provide(_kwargs: dict[str, Any]) -> dict[str, str]:
+        return {
+            "Authorization": f"Bearer {get_token()}",
+        }
+
+    return provide
+
+
 async def main() -> None:
-    """Example showing how to use a single Foundry toolbox with FoundryChatClient."""
-    print("=== Foundry Chat Client with Toolbox Example ===")
-
-    # For authentication, run `az login` in your terminal or replace
-    # AzureCliCredential with your preferred authentication option.
-    client = FoundryChatClient(
-        credential=AzureCliCredential(),
-        project_endpoint=os.environ["FOUNDRY_PROJECT_ENDPOINT"],
-        model=os.environ["FOUNDRY_MODEL"],
-    )
+    credential = DefaultAzureCredential()
 
     # Comment out if the toolbox already exists in your Foundry project.
     create_sample_toolbox(TOOLBOX_NAME)
 
-    # Omit ``version`` to resolve the toolbox's current default version at runtime.
-    toolbox = await client.get_toolbox(TOOLBOX_NAME)
-    print(f"Loaded toolbox {toolbox.name}@{toolbox.version} ({len(toolbox.tools)} tool(s))")
-
-    agent = Agent(
-        client=client,
-        instructions="You are a research assistant. Use the available tools to answer questions.",
-        tools=toolbox,
+    toolbox_tool = MCPStreamableHTTPTool(
+        name="foundry_toolbox",
+        description="Tools exposed by the configured Foundry toolbox",
+        url=os.environ["FOUNDRY_TOOLBOX_ENDPOINT"],
+        header_provider=make_toolbox_header_provider(credential),
+        load_prompts=False,
     )
 
-    print(f"User: {QUERY}")
-    result = await agent.run(QUERY)
-    print(f"Result: {result}\n")
-
-
-async def combine_toolboxes() -> None:
-    """Alternative flow: combine the tools from multiple Foundry toolboxes."""
-    client = FoundryChatClient(
-        credential=AzureCliCredential(),
-        project_endpoint=os.environ["FOUNDRY_PROJECT_ENDPOINT"],
-        model=os.environ["FOUNDRY_MODEL"],
-    )
-
-    # Comment out if the toolboxes already exist in your Foundry project.
-    create_sample_toolbox(TOOLBOX_NAME)
-    create_sample_toolbox(SECOND_TOOLBOX_NAME)
-
-    toolbox_a = await client.get_toolbox(TOOLBOX_NAME, version=TOOLBOX_VERSION)
-    toolbox_b = await client.get_toolbox(SECOND_TOOLBOX_NAME, version=SECOND_TOOLBOX_VERSION)
-    print(
-        "Loaded toolboxes: "
-        f"{toolbox_a.name}@{toolbox_a.version} ({len(toolbox_a.tools)} tool(s)), "
-        f"{toolbox_b.name}@{toolbox_b.version} ({len(toolbox_b.tools)} tool(s))"
-    )
-
-    agent = Agent(
-        client=client,
-        instructions="You are a research assistant. Use all available tools to answer questions.",
-        tools=[toolbox_a, toolbox_b],
-    )
-
-    print(f"User: {QUERY}")
-    result = await agent.run(QUERY)
-    print(f"Combined-toolbox result: {result}\n")
-
-
-async def select_tools_from_toolbox() -> None:
-    """Alternative flow: keep only a subset of toolbox tools before agent creation."""
-    client = FoundryChatClient(
-        credential=AzureCliCredential(),
-        project_endpoint=os.environ["FOUNDRY_PROJECT_ENDPOINT"],
-        model=os.environ["FOUNDRY_MODEL"],
-    )
-
-    # Comment out if the toolbox already exists in your Foundry project.
-    create_sample_toolbox(TOOLBOX_NAME)
-
-    toolbox = await client.get_toolbox(TOOLBOX_NAME, version=TOOLBOX_VERSION)
-    print(f"Loaded toolbox {toolbox.name}@{toolbox.version} ({len(toolbox.tools)} tool(s))")
-
-    selected_tools = select_toolbox_tools(
-        toolbox,
-        include_types=["code_interpreter", "mcp"],
-    )
-    print(f"Selected {len(selected_tools)} toolbox tools for the agent")
-
-    agent = Agent(
-        client=client,
-        instructions="You are a research assistant. Use only the selected toolbox tools.",
-        tools=selected_tools,
-    )
-
-    print(f"User: {QUERY}")
-    result = await agent.run(QUERY)
-    print(f"Selected-toolbox result: {result}\n")
+    async with Agent(
+        client=FoundryChatClient(
+            project_endpoint=os.environ["FOUNDRY_PROJECT_ENDPOINT"],
+            model=os.environ["FOUNDRY_MODEL"],
+            credential=credential,
+        ),
+        instructions="You are a helpful assistant. Use the available toolbox tools to answer the user.",
+        tools=toolbox_tool,
+    ) as agent:
+        query = "What tools do you have access to?"
+        print(f"User: {query}")
+        result = await agent.run(query)
+        print(f"Assistant: {result}")
 
 
 if __name__ == "__main__":
     asyncio.run(main())
-    # asyncio.run(combine_toolboxes())
-    # asyncio.run(select_tools_from_toolbox())
